+++ r-package/tests/testthat/test_read_access.R
@@ -5,8 +5,13 @@ context("read_access")
   test_that("read_access expected behavior", {
 
     # whole file
-    testthat::expect_output(object = read_access(city='nat', year=2019))
-    testthat::expect_output(object = read_access(city='nat', geometry = TRUE))
+    # testthat::expect_output(object = read_access(city='nat', year=2019))
+    # testthat::expect_output(object = read_access(city='nat', geometry = TRUE))
+
+    expect_true(is(  read_access(city='nat', geometry = TRUE), 'sf'))
+    expect_true(is(  read_access(city='nat', geometry = FALSE), 'data.frame'))
+    expect_true(is(  read_access(city='nat', geometry = FALSE, peak=FALSE), 'data.frame'))
+    expect_true(is(  read_access(city='nat', geometry = FALSE, peak=FALSE, showProgress = FALSE), 'data.frame'))
 
   })