+++ r-package/tests/testthat/test_read_grid.R
@@ -5,9 +5,7 @@ context("read_grid")
 test_that("read_grid expected behavior", {
 
   # whole file
-  testthat::expect_output(object = read_grid(city='nat'))
-  # testthat::expect_output(object = read_grid(city='nat', showProgress = FALSE))
-  # testthat::expect_output(object = read_grid(city=c('all')))
+  expect_true(is(  read_grid(city='nat'), 'sf'))
 
 })