+++ r-package/R/read_access.R
@@ -93,9 +93,14 @@ read_access <- function(city, mode = 'walk', peak = TRUE, year = 2019, geometry 
   aop_access <- download_data(file_url, progress_bar = showProgress)
 
   # peak Vs off-peak
-  cities_with_pt <- (city %in% c('for', 'rec', 'bho', 'rio', 'spo', 'cur', 'poa') |
-                    city %in% c('Fortaleza', 'Recife', 'Belo Horizonte',
-                                'Rio de Janeiro', 'S\u00e3o Paulo', 'Curitiba', 'Porto Alegre'))
+  city <- tolower(city)
+  city <- rm_accent(city)
+
+  cities_with_pt <- ( all(city %in% c('for', 'rec', 'bho', 'rio', 'spo', 'cur', 'poa')) |
+                    all(city %in% c('fortaleza', 'recife', 'belo horizonte',
+                                'rio de janeiro', 's\u00e3o paulo', 'sao paulo', 'curitiba', 'porto alegre')) )
+
+  if (isFALSE(cities_with_pt) & mode == 'public_transport') {stop("One of the selected cities does not have public transport data for that year.")}
 
   if(peak==FALSE & mode == 'public_transport' & cities_with_pt){
                   aop_access <- subset(aop_access, peak == 0)