+++ r-package/tests/testthat/test_read_grid.R
@@ -6,6 +6,7 @@ test_that("read_grid expected behavior", {
 
   # whole file
   expect_true(is(  read_grid(city='nat'), 'sf'))
+  expect_true(is(  read_grid(city='Natal'), 'sf'))
 
 })
 
@@ -17,10 +18,14 @@ test_that("read_grid errors and messages", {
   # Wrong city
   testthat::expect_error(read_grid(city = 'abcdefg'))
   testthat::expect_error(read_grid(city = 'abc'))
+  testthat::expect_error(read_grid(city = 'a'))
   testthat::expect_error(read_grid(city = 1))
   testthat::expect_error(read_grid(city = NULL))
+  testthat::expect_error(read_grid())
 
   # Wrong geometry or showProgress
-  testthat::expect_error(read_access(read_grid = 'nat', showProgress = 'aaa'))
+  testthat::expect_error( read_grid(city = 'nat', showProgress = 'aaa'))
+  testthat::expect_error( read_grid(city = 'nat', showProgress = 333))
+  testthat::expect_error( read_grid(city = 'nat', showProgress = NULL))
 
 })